Finished Listing students with more than n UC's

+++ src/Slot.h
@@ -9,7 +9,7 @@ private:
     std::string type, day;
 public:
     Slot();
-    Slot(std::string day, double startHour, double duration, std::string type);
+    Slot(std::string day, float startHour, float duration, std::string type);
     std::string get_day() const;
     double get_startHour() const;
     double get_duration() const;
